+++ lib/hephaestus_ecto/migrations/postgres.ex
@@ -4,7 +4,7 @@ defmodule HephaestusEcto.Migrations.Postgres do
   use Ecto.Migration
 
   @initial_version 1
-  @current_version 1
+  @current_version 2
   @default_prefix "public"
 
   @valid_prefix ~r/^[a-z_][a-z0-9_]*$/